+++ src/main/java/de/adito/aditoweb/timeline/definition/ITimelineConstants.java
@@ -4,12 +4,14 @@ import de.adito.aditoweb.timeline.timing.ITimelineBezier;
 import de.adito.aditoweb.timeline.timing.TimelineBeziers;
 
 /**
+ * Konstanten für die Timeline
+ *
  * @author k.mifka, 10.04.2018
  */
 public interface ITimelineConstants
 {
-    int DEFAULT_DURATION = 1000;
-    int DEFAULT_TICK_DELAY = 1;
+  long DEFAULT_DURATION = 1000;
+  long DEFAULT_TICK_DELAY = 1;
 
-    ITimelineBezier DEFAULT_BEZIER = TimelineBeziers.LINEAR;
+  ITimelineBezier DEFAULT_BEZIER = TimelineBeziers.LINEAR;
 }
